requirements.txt
dotenv
huggingface-hub
Pillow